components/studio/requirements.txt
scikit-learn
minio
certifi==2019.9.11
chardet==3.0.4
Django==2.2.6
django-appconf==1.0.3
django-compressor==2.3
django-cors-middleware==1.4.0
django-filter==2.2.0
django-oauth-toolkit==1.2.0
djangorestframework==3.10.3
idna==2.8
oauthlib==3.1.0
psycopg2-binary==2.8.4
pytz==2019.3
rcssmin==1.0.6
requests==2.22.0
rjsmin==1.1.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
vine==1.2.0
gitdb2==2.0.6
GitPython==3.0.5
amqp==2.4.2
billiard==3.5.0.5
celery==4.2.1
kombu==4.4.0
redis==3.2.0
websocket-client==0.56.0
kubernetes==10.0.1
PyYAML
numpy==1.16.1
matplotlib==3.0.3
django-bootstrap-modal-forms==1.5.0
django-yamlfield==1.0.3
